requirements.txt
Flask==2.3.3
googletrans==4.0.0rc1
langdetect==1.0.9
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
python-dotenv==1.0.0